targets

+++ UNet/train_unet.py
@@ -61,9 +61,8 @@ def calculate_mAP(preds, targets, threshold=0.5):
     
     #preds_bin_flatten = torch.flatten(preds_bin)
     #targets_flatten = torch.flatten(targets)
-    targets = targets.long()
     
-    ap = average_precision(preds, targets, task="binary")
+    ap = average_precision(preds, targets.long(), task="binary")
     return ap
 
 def calculate_accuracy(preds, targets):
